Vk Api update version. Vk dialog is broken

+++ app/src/main/java/com/example/korol/onechatapp/logic/vk/json/VkDialogJsonParser.java
@@ -21,7 +21,7 @@ public class VkDialogJsonParser extends AsyncOperation<String, List<IMessage>> {
         JSONArray allMessages = new JSONObject(json).getJSONArray("response");
         for (int i = 1; i < allMessages.length(); i++) {
             JSONObject jsonMessageObject = allMessages.getJSONObject(i);
-            VkMessage.Builder builder = new VkMessage.Builder().setText(jsonMessageObject.getString("body")).setSender(VkIdToUserStorage.getUser(jsonMessageObject.getInt("from_id"))).setDate(new Date(jsonMessageObject.getLong("date") * 1000));
+            VkMessage.Builder builder = new VkMessage.Builder().setText(jsonMessageObject.getString("body")).setSender(VkIdToUserStorage.getUser(jsonMessageObject.getLong("from_id"))).setDate(new Date(jsonMessageObject.getLong("date") * 1000));
             list.add(builder.build());
         }
         return list;
